+++ python/lsst/daf/butler/core/_butlerUri/s3.py
@@ -47,6 +47,40 @@ if TYPE_CHECKING:
         pass
     from ..datastore import DatastoreTransaction
 
+# https://pypi.org/project/backoff/
+try:
+    import backoff
+except ImportError:
+    class Backoff():
+        @staticmethod
+        def expo(func: Callable, *args: Any, **kwargs: Any) -> Callable:
+            return func
+
+        @staticmethod
+        def on_exception(func: Callable, *args: Any, **kwargs: Any) -> Callable:
+            return func
+
+    backoff = Backoff
+
+# settings for "backoff" retry decorators. these retries are belt-and-
+# suspenders along with the retries built into Boto3, to account for
+# semantic differences in errors between S3-like providers.
+retryable_io_errors = (
+    # http.client
+    ImproperConnectionState, HTTPException,
+    # urllib3.exceptions
+    RequestError, HTTPError,
+    # built-ins
+    TimeoutError, ConnectionError)
+retryable_client_errors = (
+    # botocore.exceptions
+    ClientError,
+    # built-ins
+    PermissionError)
+all_retryable_errors = retryable_client_errors + retryable_io_errors
+max_retry_time = 60
+
+
 log = logging.getLogger(__name__)
